Move `BaseDirectories` types into its own module
The public types are re-exported in the crate root again.
